+++ src/mnelab/mainwindow.py
@@ -751,13 +751,12 @@ class MainWindow(QMainWindow):
         montages = natural_sort(mne.channels.get_builtin_montages())
         dialog = MontageDialog(self, montages)
         if dialog.exec():
-            name = dialog.montages.selectedItems()[0].data(0)
-            montage = mne.channels.make_standard_montage(name)
+            montage = dialog.montage
             ch_names = self.model.current["data"].info["ch_names"]
             # check if at least one channel name matches a name in the montage
-            if set(ch_names) & set(montage.ch_names):
+            if set(ch_names) & set(montage.montage.ch_names):
                 self.model.set_montage(
-                    name,
+                    montage,
                     match_case=dialog.match_case.isChecked(),
                     match_alias=dialog.match_alias.isChecked(),
                     on_missing="ignore"